Assembly: Create SoSwitchMarker.py to externalize the coin stuff that draw the joint's markers(JCS)
